Add replace command
Reads a line from stdin and parses it into a non-standard dep type. For
now supports a constraint of exactly `local`. Prints a string which is a
valid pubspec replacement for the dependency.

If the package name starts with `#` it is stripped to support running
from source within this package, otherwise a constraint like
`package: local` would cause failures in pub.