Add support for git dependencies
The package must be checked out locally and have at least one remote
that points to the dart-lang or google org on github. Assumes that the
url is a `git://`, `https://`, or `git@`. Always returns a `git@`.

Local repository remote information is scraped by running the `git`
command and parsing the output.